Moves chat sending features to proper chat form input component
